Change key bindings behaviors for ENTER, W and I
Changed the behavior when ENTER is pressed, so there isn't the automatic
click event, which opens word editing. This allows using W and I to
quickly mark the unknown word as well-known or ignored.

This introduced a bug where if ENTER is pressed before invoking some
kind of tooltip through overlib.js library, there will be an error thrown
in the console every time keys like W or I are pressed, until some kind
of tooltip is invoked. I haven't found a way to prevent this yet.