Make mouse-less adding of words easier
Now you can keep pressing ENTER, and W if known, I if ignore,
and newly E if you want to specify translation and status of the word.
When you press E, start typing in the translation, and when finished,
press ENTER to set status to 1, or Tab and change to appropriate status
with arrows. Then press ENTER.